test(app): keep E2E fixture helpers bounded
