+++ tests/_helpers/apps.ts
@@ -8,6 +8,16 @@ import {
   ensureLocalVizePackagesBuilt,
   ensureSymlink,
 } from "./vize-local-packages.ts";
+import {
+  FRONTEND_PHPCON_E2E_API_BASE,
+  FRONTEND_PHPCON_E2E_ENV,
+  NPMX_E2E_ENV,
+  VUEFES_E2E_ENV,
+  execNpxCommand,
+  npmxGeneratorTaskArgs,
+  readDotenvValue,
+  writeFrontendPhpconStaffRoute,
+} from "./app-fixture-runtime.ts";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -55,26 +65,6 @@ export interface AppConfig {
 
 // --- Setup helpers ---
 const MISSKEY_FLUENT_EMOJI_RE = /\/fluent-emoji(?:s)?\/([0-9a-z-]+\.png)\b/g;
-const NPMX_E2E_ENV = {
-  NUXT_SESSION_PASSWORD: "e2e-test-dummy-session-password-32chars!",
-  VIZE_E2E_DISABLE_LUNARIA: "1",
-} as const;
-export const FRONTEND_PHPCON_E2E_API_BASE = "/__vize_e2e/api";
-export const FRONTEND_PHPCON_STAFF_ROUTE_RELATIVE_PATH = path.join(
-  "server",
-  "routes",
-  "__vize_e2e",
-  "api",
-  "staff.get.ts",
-);
-const VUEFES_E2E_ENV = {
-  AUTH_SECRET: "e2e-test-dummy-auth-secret-32chars!",
-  NEXTAUTH_SECRET: "e2e-test-dummy-auth-secret-32chars!",
-} as const;
-const FRONTEND_PHPCON_E2E_ENV = {
-  NUXT_PUBLIC_API_BASE: FRONTEND_PHPCON_E2E_API_BASE,
-  NUXT_TELEMETRY_DISABLED: "1",
-} as const;
 const VUE_BETA_OVERRIDES = {
   "@vue/compiler-core": "3.6.0-beta.10",
   "@vue/compiler-dom": "3.6.0-beta.10",
@@ -92,49 +82,10 @@ const TRANSPARENT_PNG = Buffer.from(
   "base64",
 );
 
-function readDotenvValue(filePath: string, key: string): string | undefined {
-  if (!fs.existsSync(filePath)) {
-    return undefined;
-  }
-
-  const prefix = `${key}=`;
-  for (const line of fs.readFileSync(filePath, "utf-8").split(/\r?\n/)) {
-    if (!line.startsWith(prefix)) {
-      continue;
-    }
-    const value = line.slice(prefix.length);
-    const quote = value.at(0);
-    if ((quote === "'" || quote === '"') && value.endsWith(quote)) {
-      return value.slice(1, -1);
-    }
-    return value;
-  }
-  return undefined;
-}
-
 interface PnpmInstallOptions {
   env?: NodeJS.ProcessEnv;
   ignoreScripts?: boolean;
   timeout?: number;
-}
-
-function execNpxCommand(
-  args: string[],
-  opts: { cwd: string; env?: NodeJS.ProcessEnv; timeout?: number },
-): void {
-  const executable = process.platform === "win32" ? (process.env.ComSpec ?? "cmd.exe") : "npx";
-  const executableArgs =
-    process.platform === "win32" ? ["/d", "/s", "/c", "npx.cmd", ...args] : args;
-  execFileSync(executable, executableArgs, {
-    cwd: opts.cwd,
-    env: opts.env,
-    stdio: "inherit",
-    timeout: opts.timeout ?? 600_000,
-  });
-}
-
-export function npmxGeneratorTaskArgs(task: "generate:lexicons" | "generate:sprite"): string[] {
-  return ["-y", "pnpm@10", "exec", "vp", "run", task];
 }
 
 export function installPnpmDependencies(cwd: string, options: PnpmInstallOptions = {}): void {
@@ -1470,41 +1421,7 @@ function patchFrontendPhpconVisualFixture(frontendDir: string): void {
 }));
 `,
   );
-  ensureFileContent(
-    path.join(frontendDir, FRONTEND_PHPCON_STAFF_ROUTE_RELATIVE_PATH),
-    `export default defineEventHandler(() => ({
-  staff_types: [
-    {
-      name: "実行委員長",
-      name_en: "Chair",
-      staff: [
-        {
-          id: "chair-1",
-          name: "Hokkaido Chair",
-          url: "https://example.com/staff/chair",
-        },
-      ],
-    },
-    {
-      name: "コアスタッフ",
-      name_en: "Core Staff",
-      staff: [
-        {
-          id: "core-1",
-          name: "Core Staff A",
-          url: "https://example.com/staff/core-a",
-        },
-        {
-          id: "core-2",
-          name: "Core Staff B",
-          url: "https://example.com/staff/core-b",
-        },
-      ],
-    },
-  ],
-}));
-`,
-  );
+  writeFrontendPhpconStaffRoute(frontendDir, ensureFileContent);
 
   for (const imageDir of [
     path.join(frontendDir, "public", "individual-sponsors"),
